+++ CHANGELOG.md
@@ -18,6 +18,7 @@
           `tests/` subdirectory.
         * Rename `{Basic,Extended}XxxTest` to `Xxx{Basic,Extended}Test`.
     * Rename `AceTimePython` library to `acetimepy`.
+    * Rename `AceTimeC` library to `acetimec`.
 * 1.5.2 (2023-04-01, TZDB 2023c)
     * Upgrade TZDB from 2023b to 2023c.
         * https://mm.icann.org/pipermail/tz-announce/2023-March/000079.html